major cabal completion changes
complete `lib:`, `test:`, `bench:` as before;
complete <package>:<component>. If there are multiple component of the
same kind the <component> part will also include one of `libs`, `tests`,
`benches`, exes`.  For the common case of a single library it can be
picked by `lib:libname`, `libname:libname` is not included (should it?).